Insert or Remove Named Elements

Insert elements from y into x by name, or remove elements from x by name. Works for vectors, lists, environments and data frames and data tables. Objects with reference semantic (environment() and data.table::data.table()) might be modified in-place.

Usage

insert_named(x, y)

# S3 method for class '`NULL`'
insert_named(x, y)

# Default S3 method
insert_named(x, y)

# S3 method for class 'environment'
insert_named(x, y)

# S3 method for class 'data.frame'
insert_named(x, y)

# S3 method for class 'data.table'
insert_named(x, y)

remove_named(x, nn)

# S3 method for class 'environment'
remove_named(x, nn)

# S3 method for class 'data.frame'
remove_named(x, nn)

# S3 method for class 'data.table'
remove_named(x, nn)

Arguments

x

(vector() | list() | environment() | data.table::data.table())
Object to insert elements into, or remove elements from. Changes are by-reference for environments and data tables.

y

(list())
List of elements to insert into x.

nn

(character())
Character vector of elements to remove.

Value

Modified object.

Examples

x = list(a = 1, b = 2)
insert_named(x, list(b = 3, c = 4))
#> $a
#> [1] 1
#> 
#> $b
#> [1] 3
#> 
#> $c
#> [1] 4
#> 
remove_named(x, "b")
#> $a
#> [1] 1
#>